Add gpu_arch to filter and enable 64-bit asum tests (#761)
[ROCm/hipBLAS commit: 3ff76182dd90d8f2ca30ff8f3bfe8f5bbf96c446]

+++ clients/include/unit.h
@@ -1,5 +1,5 @@
 /* ************************************************************************
- * Copyright (C) 2016-2022 Advanced Micro Devices, Inc. All rights reserved.
+ * Copyright (C) 2016-2024 Advanced Micro Devices, Inc. All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -50,18 +50,27 @@
 // case
 // a wrapper will cause the loop keep going
 template <typename T>
-void unit_check_general(int M, int N, int lda, T* hCPU, T* hGPU);
+void unit_check_general(int64_t M, int64_t N, int64_t lda, T* hCPU, T* hGPU);
 
 template <typename T>
-void unit_check_general(
-    int M, int N, int batch_count, int lda, hipblasStride stride_A, T* hCPU, T* hGPU);
+void unit_check_general(int64_t       M,
+                        int64_t       N,
+                        int64_t       batch_count,
+                        int64_t       lda,
+                        hipblasStride stride_A,
+                        T*            hCPU,
+                        T*            hGPU);
 
 template <typename T>
-void unit_check_general(int M, int N, int batch_count, int lda, T** hCPU, T** hGPU);
+void unit_check_general(int64_t M, int64_t N, int64_t batch_count, int64_t lda, T** hCPU, T** hGPU);
 
 template <typename T>
-void unit_check_general(
-    int M, int N, int batch_count, int lda, host_vector<T> hCPU[], host_vector<T> hGPU[]);
+void unit_check_general(int64_t        M,
+                        int64_t        N,
+                        int64_t        batch_count,
+                        int64_t        lda,
+                        host_vector<T> hCPU[],
+                        host_vector<T> hGPU[]);
 
 template <typename T>
 void unit_check_error(T error, T tolerance)
